HACK:domd: Lower USERADD_ERROR_DYNAMIC severity
Systemd-networkd introduction also adds a user for it. That user is not
described in AGL's native passwd. So we got an error at the moment of recipes
dependencies resolving due to  AGL's default setup to error on a user which
ID is not set static.
Currently it was not found a reasonable way to append a user to AGL's passwd
file, so we lower the severity of the issue down to warning for systemd recipe.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>